Feat: - extract functions from HomeActivity.kt that do not correspond to the view.
